+++ database/migrations/2024_06_09_150920_add_activation_token_to_users_table.php
@@ -4,25 +4,25 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class AddActivationTokenToUsersTable extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('activation_token')->nullable();
+            
+            if (!Schema::hasColumn('users', 'activation_token')) {
+                $table->string('activation_token')->nullable()->after('password');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('activation_token');
+            if (Schema::hasColumn('users', 'activation_token')) {
+                $table->dropColumn('activation_token');
+            }
         });
     }
-};
+}
+
